Update env2cfg.sh
This simplified script uses a for loop to iterate through the environment variables and checks if they are set. If the variable is set, the script checks if it exists in the app.cfg file and updates or adds the variable accordingly.

The cp command now redirects the output to /dev/null to avoid printing the "Permission denied" message when the user doesn't have permission to write to the backup file.

The &>/dev/null redirection is used to hide the "Permission denied" message when creating the backup file.

The grep -q option is used to avoid printing the matched lines when searching for the variable in the app.cfg file.

The ${!var} syntax is used to expand the value of the variable whose name is stored in the variable var.