Update env2cfg.sh
Lets try this.

This code uses an array to store the environment variables and their corresponding config names. It then loops through the array, checks if the environment variable is set, and updates the config file accordingly.

This refactored code is more concise and easier to maintain. If you need to add or remove variables, you only need to modify the variables array.